Check the symlink is not already there.

+++ install.sh
@@ -1,4 +1,6 @@
 # Following two lines are needed for the Custom CSS Loader extension
 sudo chown -R $(whoami) "$(which code)"
 sudo chown -R $(whoami) /usr/share/code
-ln -s ~/Projects/vscode-pksthemes ~/.vscode/extensions/vscode-pksthemes
+if [ ! -L ~/.vscode/extensions/vscode-pksthemes ]; then
+  ln -s ~/Projects/vscode-pksthemes ~/.vscode/extensions/vscode-pksthemes
+fi
